@@ -1,5 +1,5 @@
 class Admin::AdminController < ApplicationController
-  before_filter :check_admin_access
+  before_action :check_admin_access
 
   private
   def check_admin_access
